Minor bug fixes

+++ workers/email_worker.py
@@ -18,26 +18,21 @@ class EmailWorker:
         message_id = message["message_id"]
         deletion_id = message["deletion_id"]
         unique_id = message["unique_id"]
-        print("MESSAGE ID:", message_id)
-        print("QUEUE ID:", unique_id)
         email = extract_information_from_email(message_id)
         conv_id = email["conversationId"]
         body = email["body"]
-        print("ckecking if conv alrerdy exists,....")
         answer = self.sql_worker.get_conversation_info(conv_id)
         if answer == None:
-            print("creating ticket")
             agent = self.routing_service.get_next_available_agent()
+            if agent is None:
+                raise RuntimeError("No agent is available right now")
             data = self.convert_email_to_ticket(email, agent)
             response = self.ticketing_service.create_ticket(data)
             ticket_id = response["IDs"][0]
             self.ticketing_service.append_comment(3, ticket_id, body)
             self.sql_worker.create_conversation(conv_id, agent.unique_id, ticket_id)
             self.queue_client.delete_message(unique_id, deletion_id)
-            print("deleted")
         else:
-            print("appending ticket")
-            agent_id = answer["agent_id"]
             ticket_id = answer["ticket_id"]
             self.ticketing_service.append_comment(3, ticket_id, body)
             self.queue_client.delete_message(unique_id, deletion_id)
